+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/Role.py
