+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/SAMLRoleMapping.py
