+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/SMTPSetting.py
