+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/SystemLocale.py
