+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/Team.py
