+++ CheckmarxPythonSDK/CxAccessControl/accesscontrol/dto/WindowsDomain.py
