+++ CheckmarxPythonSDK/utilities/CxError.py
