+++ CheckmarxPythonSDK/utilities/compat.py
