+++ CheckmarxPythonSDK/utilities/configUtility.py
